refactor: implement new generic and extension functions in the use cases
