+++ domain/src/main/java/com/example/domain/useCase/GetTvShowByKeywordUseCase.kt
@@ -1,9 +1,12 @@
 package com.example.domain.useCase
 
+import com.example.domain.common.ContentFilteringExtensions.filterByCategory
+import com.example.domain.common.ContentFilteringExtensions.filterByMinRating
+import com.example.domain.common.ContentFilteringExtensions.sortByPopularityDescending
+import com.example.domain.common.ContentFilteringExtensions.throwIfEmpty
 import com.example.domain.exceptions.NoSearchByKeywordResultFoundException
 import com.example.domain.repository.TvShowRepository
 import com.example.entity.TvShow
-import kotlin.math.roundToInt
 
 class GetTvShowByKeywordUseCase(
     private val tvShowRepository: TvShowRepository
@@ -11,20 +14,13 @@ class GetTvShowByKeywordUseCase(
 
     suspend operator fun invoke(
         keyword: String,
-        rating: Float = 0f,
+        rating: Int = 0,
         tvShowGenreId: Int = 0
     ): List<TvShow> {
         return tvShowRepository.getTvShowByKeyword(keyword = keyword)
-            .filter { tv -> tv.rating.roundToInt() >= rating }
-            .filter { tvGenre ->
-                if (tvShowGenreId == 0)
-                    return@filter true
-
-                tvGenre.categories.any { it.id == tvShowGenreId.toLong() }
-            }
-            .sortedByDescending { it.popularity }
-            .ifEmpty {
-                throw NoSearchByKeywordResultFoundException()
-            }
+            .filterByMinRating(rating)
+            .filterByCategory(tvShowGenreId)
+            .sortByPopularityDescending()
+            .throwIfEmpty { NoSearchByKeywordResultFoundException() }
     }
 }
